+++ ChangeLog.md
@@ -12,6 +12,9 @@ Behavior changes:
 Other enhancements:
 
 Bug fixes:
+* When a package contained sublibraries, stack was always recompiling the
+  package. This has been fixed now, no recompilation is being done because of
+  sublibraries. See [#3899](https://github.com/commercialhaskell/stack/issues/3899).
 
 
 ## v1.7.0.1 (releases candidate)